Move all component protobuf stuff into Components.proto
