--- proto/streamlit/proto/ArrowDataframe.proto
@@ -1,25 +1,0 @@
-/**
- * Copyright 2018-2021 Streamlit Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
-*/
-
-syntax = "proto3";
-
-import "streamlit/proto/ArrowTable.proto";
-
-message ArrowDataframe {
-  ArrowTable data = 1;
-  uint32 height = 2;
-  uint32 width = 3;
-}
