--- proto/streamlit/proto/ArrowTable.proto
@@ -1,31 +1,0 @@
-/**
- * Copyright 2018-2021 Streamlit Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
-*/
-
-syntax = "proto3";
-
-message ArrowTable {
-  bytes data = 1;
-  bytes index = 2;
-  bytes columns = 3;
-  Styler styler = 5;
-}
-
-message Styler {
-  string uuid = 1;
-  string caption = 2;
-  string styles = 3;
-  bytes display_values = 4;
-}
